refactor: clean up the code after rebase
