--- src/components/__tests__/utils/index.js
@@ -1,1 +1,0 @@
-export * from './utils';
